Split benchmark code into mods
